+++ Tests/SentryTests/SentryOptionsTest.m
@@ -87,6 +87,13 @@
         infoDict[@"CFBundleShortVersionString"], infoDict[@"CFBundleVersion"]];
 }
 
+#if TARGET_OS_OSX
+- (void)testEnableReportingUncaughtExceptions
+{
+    [self testBooleanField:@"enableReportingUncaughtExceptions" defaultValue:NO];
+}
+#endif // TARGET_OS_OSX
+
 - (void)testEnvironment
 {
     SentryOptions *options = [self getValidOptions:@{}];