+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -180,6 +180,7 @@
 #import "SentryScreenshotIntegration.h"
 #import "SentrySwift.h"
 #import "SentrySwiftAsyncIntegration.h"
+#import "SentryUncaughtNSExceptions.h"
 
 #import "Sentry/Sentry-Swift.h"
 #import "SentryBinaryImageCache+Private.h"